requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
sqladmin==0.18.0

# Authentication
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.6
itsdangerous==2.1.2

# Celery
celery==5.3.4
redis==5.0.1
flower==2.0.1

# HTTP client
httpx==0.25.2
aiohttp==3.12.14
requests==2.32.4
# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1
PyJWT==2.10.1


aiogram>=3.0.0

streamlit
openapi-python-client
yookassa
